move everything about pids to apps_proc_pids.c
