+++ src/collectors/apps.plugin/apps_plugin.c
@@ -96,30 +96,6 @@ NETDATA_DOUBLE
     cmajflt_fix_ratio = 1.0;
 
 // ----------------------------------------------------------------------------
-
-#ifdef __FreeBSD__
-#include <sys/user.h>
-#endif
-
-// ----------------------------------------------------------------------------
-// per O/S configuration
-
-// the minimum PID of the system
-// this is also the pid of the init process
-#define INIT_PID 1
-
-// if the way apps.plugin will work, will read the entire process list,
-// including the resource utilization of each process, instantly
-// set this to 1
-// when set to 0, apps.plugin builds a sort list of processes, in order
-// to process children processes, before parent processes
-#ifdef __FreeBSD__
-#define ALL_PIDS_ARE_READ_INSTANTLY 1
-#else
-#define ALL_PIDS_ARE_READ_INSTANTLY 0
-#endif
-
-// ----------------------------------------------------------------------------
 // factor for calculating correct CPU time values depending on units of raw data
 unsigned int time_factor = 0;
 
@@ -166,8 +142,7 @@ struct pid_stat
 // read parents before children. This is needed to prevent a situation where
 // a child is found running, but until we read its parent, it has exited and
 // its parent has accumulated its resources.
-static pid_t
-        *all_pids_sortlist = NULL;
+pid_t *all_pids_sortlist = NULL;
 #endif
 
 // ----------------------------------------------------------------------------
@@ -236,451 +211,6 @@ int managed_log(struct pid_stat *p, PID_LOG log, int status) {
     }
 
     return status;
-}
-
-// ----------------------------------------------------------------------------
-
-static inline int debug_print_process_and_parents(struct pid_stat *p, usec_t time) {
-    char *prefix = "\\_ ";
-    int indent = 0;
-
-    if(p->parent)
-        indent = debug_print_process_and_parents(p->parent, p->stat_collected_usec);
-    else
-        prefix = " > ";
-
-    char buffer[indent + 1];
-    int i;
-
-    for(i = 0; i < indent ;i++) buffer[i] = ' ';
-    buffer[i] = '\0';
-
-    fprintf(stderr, "  %s %s%s (%d %s %"PRIu64""
-        , buffer
-        , prefix
-        , p->comm
-        , p->pid
-        , p->updated?"running":"exited"
-        , p->stat_collected_usec - time
-        );
-
-    if(p->utime)   fprintf(stderr, " utime=" KERNEL_UINT_FORMAT,   p->utime);
-    if(p->stime)   fprintf(stderr, " stime=" KERNEL_UINT_FORMAT,   p->stime);
-    if(p->gtime)   fprintf(stderr, " gtime=" KERNEL_UINT_FORMAT,   p->gtime);
-    if(p->cutime)  fprintf(stderr, " cutime=" KERNEL_UINT_FORMAT,  p->cutime);
-    if(p->cstime)  fprintf(stderr, " cstime=" KERNEL_UINT_FORMAT,  p->cstime);
-    if(p->cgtime)  fprintf(stderr, " cgtime=" KERNEL_UINT_FORMAT,  p->cgtime);
-    if(p->minflt)  fprintf(stderr, " minflt=" KERNEL_UINT_FORMAT,  p->minflt);
-    if(p->cminflt) fprintf(stderr, " cminflt=" KERNEL_UINT_FORMAT, p->cminflt);
-    if(p->majflt)  fprintf(stderr, " majflt=" KERNEL_UINT_FORMAT,  p->majflt);
-    if(p->cmajflt) fprintf(stderr, " cmajflt=" KERNEL_UINT_FORMAT, p->cmajflt);
-    fprintf(stderr, ")\n");
-
-    return indent + 1;
-}
-
-static inline void debug_print_process_tree(struct pid_stat *p, char *msg __maybe_unused) {
-    debug_log("%s: process %s (%d, %s) with parents:", msg, p->comm, p->pid, p->updated?"running":"exited");
-    debug_print_process_and_parents(p, p->stat_collected_usec);
-}
-
-static inline void debug_find_lost_child(struct pid_stat *pe, kernel_uint_t lost, int type) {
-    int found = 0;
-    struct pid_stat *p = NULL;
-
-    for(p = root_of_pids; p ; p = p->next) {
-        if(p == pe) continue;
-
-        switch(type) {
-            case 1:
-                if(p->cminflt > lost) {
-                    fprintf(stderr, " > process %d (%s) could use the lost exited child minflt " KERNEL_UINT_FORMAT " of process %d (%s)\n", p->pid, p->comm, lost, pe->pid, pe->comm);
-                    found++;
-                }
-                break;
-
-            case 2:
-                if(p->cmajflt > lost) {
-                    fprintf(stderr, " > process %d (%s) could use the lost exited child majflt " KERNEL_UINT_FORMAT " of process %d (%s)\n", p->pid, p->comm, lost, pe->pid, pe->comm);
-                    found++;
-                }
-                break;
-
-            case 3:
-                if(p->cutime > lost) {
-                    fprintf(stderr, " > process %d (%s) could use the lost exited child utime " KERNEL_UINT_FORMAT " of process %d (%s)\n", p->pid, p->comm, lost, pe->pid, pe->comm);
-                    found++;
-                }
-                break;
-
-            case 4:
-                if(p->cstime > lost) {
-                    fprintf(stderr, " > process %d (%s) could use the lost exited child stime " KERNEL_UINT_FORMAT " of process %d (%s)\n", p->pid, p->comm, lost, pe->pid, pe->comm);
-                    found++;
-                }
-                break;
-
-            case 5:
-                if(p->cgtime > lost) {
-                    fprintf(stderr, " > process %d (%s) could use the lost exited child gtime " KERNEL_UINT_FORMAT " of process %d (%s)\n", p->pid, p->comm, lost, pe->pid, pe->comm);
-                    found++;
-                }
-                break;
-        }
-    }
-
-    if(!found) {
-        switch(type) {
-            case 1:
-                fprintf(stderr, " > cannot find any process to use the lost exited child minflt " KERNEL_UINT_FORMAT " of process %d (%s)\n", lost, pe->pid, pe->comm);
-                break;
-
-            case 2:
-                fprintf(stderr, " > cannot find any process to use the lost exited child majflt " KERNEL_UINT_FORMAT " of process %d (%s)\n", lost, pe->pid, pe->comm);
-                break;
-
-            case 3:
-                fprintf(stderr, " > cannot find any process to use the lost exited child utime " KERNEL_UINT_FORMAT " of process %d (%s)\n", lost, pe->pid, pe->comm);
-                break;
-
-            case 4:
-                fprintf(stderr, " > cannot find any process to use the lost exited child stime " KERNEL_UINT_FORMAT " of process %d (%s)\n", lost, pe->pid, pe->comm);
-                break;
-
-            case 5:
-                fprintf(stderr, " > cannot find any process to use the lost exited child gtime " KERNEL_UINT_FORMAT " of process %d (%s)\n", lost, pe->pid, pe->comm);
-                break;
-        }
-    }
-}
-
-static inline kernel_uint_t remove_exited_child_from_parent(kernel_uint_t *field, kernel_uint_t *pfield) {
-    kernel_uint_t absorbed = 0;
-
-    if(*field > *pfield) {
-        absorbed += *pfield;
-        *field -= *pfield;
-        *pfield = 0;
-    }
-    else {
-        absorbed += *field;
-        *pfield -= *field;
-        *field = 0;
-    }
-
-    return absorbed;
-}
-
-static inline void process_exited_processes() {
-    struct pid_stat *p;
-
-    for(p = root_of_pids; p ; p = p->next) {
-        if(p->updated || !p->stat_collected_usec)
-            continue;
-
-        kernel_uint_t utime  = (p->utime_raw + p->cutime_raw)   * (USEC_PER_SEC * RATES_DETAIL) / (p->stat_collected_usec - p->last_stat_collected_usec);
-        kernel_uint_t stime  = (p->stime_raw + p->cstime_raw)   * (USEC_PER_SEC * RATES_DETAIL) / (p->stat_collected_usec - p->last_stat_collected_usec);
-        kernel_uint_t gtime  = (p->gtime_raw + p->cgtime_raw)   * (USEC_PER_SEC * RATES_DETAIL) / (p->stat_collected_usec - p->last_stat_collected_usec);
-        kernel_uint_t minflt = (p->minflt_raw + p->cminflt_raw) * (USEC_PER_SEC * RATES_DETAIL) / (p->stat_collected_usec - p->last_stat_collected_usec);
-        kernel_uint_t majflt = (p->majflt_raw + p->cmajflt_raw) * (USEC_PER_SEC * RATES_DETAIL) / (p->stat_collected_usec - p->last_stat_collected_usec);
-
-        if(utime + stime + gtime + minflt + majflt == 0)
-            continue;
-
-        if(unlikely(debug_enabled)) {
-            debug_log("Absorb %s (%d %s total resources: utime=" KERNEL_UINT_FORMAT " stime=" KERNEL_UINT_FORMAT " gtime=" KERNEL_UINT_FORMAT " minflt=" KERNEL_UINT_FORMAT " majflt=" KERNEL_UINT_FORMAT ")"
-                , p->comm
-                , p->pid
-                , p->updated?"running":"exited"
-                , utime
-                , stime
-                , gtime
-                , minflt
-                , majflt
-                );
-            debug_print_process_tree(p, "Searching parents");
-        }
-
-        struct pid_stat *pp;
-        for(pp = p->parent; pp ; pp = pp->parent) {
-            if(!pp->updated) continue;
-
-            kernel_uint_t absorbed;
-            absorbed = remove_exited_child_from_parent(&utime,  &pp->cutime);
-            if(unlikely(debug_enabled && absorbed))
-                debug_log(" > process %s (%d %s) absorbed " KERNEL_UINT_FORMAT " utime (remaining: " KERNEL_UINT_FORMAT ")", pp->comm, pp->pid, pp->updated?"running":"exited", absorbed, utime);
-
-            absorbed = remove_exited_child_from_parent(&stime,  &pp->cstime);
-            if(unlikely(debug_enabled && absorbed))
-                debug_log(" > process %s (%d %s) absorbed " KERNEL_UINT_FORMAT " stime (remaining: " KERNEL_UINT_FORMAT ")", pp->comm, pp->pid, pp->updated?"running":"exited", absorbed, stime);
-
-            absorbed = remove_exited_child_from_parent(&gtime,  &pp->cgtime);
-            if(unlikely(debug_enabled && absorbed))
-                debug_log(" > process %s (%d %s) absorbed " KERNEL_UINT_FORMAT " gtime (remaining: " KERNEL_UINT_FORMAT ")", pp->comm, pp->pid, pp->updated?"running":"exited", absorbed, gtime);
-
-            absorbed = remove_exited_child_from_parent(&minflt, &pp->cminflt);
-            if(unlikely(debug_enabled && absorbed))
-                debug_log(" > process %s (%d %s) absorbed " KERNEL_UINT_FORMAT " minflt (remaining: " KERNEL_UINT_FORMAT ")", pp->comm, pp->pid, pp->updated?"running":"exited", absorbed, minflt);
-
-            absorbed = remove_exited_child_from_parent(&majflt, &pp->cmajflt);
-            if(unlikely(debug_enabled && absorbed))
-                debug_log(" > process %s (%d %s) absorbed " KERNEL_UINT_FORMAT " majflt (remaining: " KERNEL_UINT_FORMAT ")", pp->comm, pp->pid, pp->updated?"running":"exited", absorbed, majflt);
-        }
-
-        if(unlikely(utime + stime + gtime + minflt + majflt > 0)) {
-            if(unlikely(debug_enabled)) {
-                if(utime) debug_find_lost_child(p, utime, 3);
-                if(stime) debug_find_lost_child(p, stime, 4);
-                if(gtime) debug_find_lost_child(p, gtime, 5);
-                if(minflt) debug_find_lost_child(p, minflt, 1);
-                if(majflt) debug_find_lost_child(p, majflt, 2);
-            }
-
-            p->keep = true;
-
-            debug_log(" > remaining resources - KEEP - for another loop: %s (%d %s total resources: utime=" KERNEL_UINT_FORMAT " stime=" KERNEL_UINT_FORMAT " gtime=" KERNEL_UINT_FORMAT " minflt=" KERNEL_UINT_FORMAT " majflt=" KERNEL_UINT_FORMAT ")"
-                , p->comm
-                , p->pid
-                , p->updated?"running":"exited"
-                , utime
-                , stime
-                , gtime
-                , minflt
-                , majflt
-            );
-
-            for(pp = p->parent; pp ; pp = pp->parent) {
-                if(pp->updated) break;
-                pp->keep = true;
-
-                debug_log(" > - KEEP - parent for another loop: %s (%d %s)"
-                    , pp->comm
-                    , pp->pid
-                    , pp->updated?"running":"exited"
-                );
-            }
-
-            p->utime_raw   = utime  * (p->stat_collected_usec - p->last_stat_collected_usec) / (USEC_PER_SEC * RATES_DETAIL);
-            p->stime_raw   = stime  * (p->stat_collected_usec - p->last_stat_collected_usec) / (USEC_PER_SEC * RATES_DETAIL);
-            p->gtime_raw   = gtime  * (p->stat_collected_usec - p->last_stat_collected_usec) / (USEC_PER_SEC * RATES_DETAIL);
-            p->minflt_raw  = minflt * (p->stat_collected_usec - p->last_stat_collected_usec) / (USEC_PER_SEC * RATES_DETAIL);
-            p->majflt_raw  = majflt * (p->stat_collected_usec - p->last_stat_collected_usec) / (USEC_PER_SEC * RATES_DETAIL);
-            p->cutime_raw = p->cstime_raw = p->cgtime_raw = p->cminflt_raw = p->cmajflt_raw = 0;
-
-            debug_log(" ");
-        }
-        else
-            debug_log(" > totally absorbed - DONE - %s (%d %s)"
-                , p->comm
-                , p->pid
-                , p->updated?"running":"exited"
-            );
-    }
-}
-
-static inline void link_all_processes_to_their_parents(void) {
-    struct pid_stat *p, *pp;
-
-    // link all children to their parents
-    // and update children count on parents
-    for(p = root_of_pids; p ; p = p->next) {
-        // for each process found
-
-        p->sortlist = 0;
-        p->parent = NULL;
-
-        if(unlikely(!p->ppid)) {
-            //unnecessary code from apps_plugin.c
-            //p->parent = NULL;
-            continue;
-        }
-
-        pp = all_pids[p->ppid];
-        if(likely(pp)) {
-            p->parent = pp;
-            pp->children_count++;
-
-            if(unlikely(debug_enabled || (p->target && p->target->debug_enabled)))
-                debug_log_int("child %d (%s, %s) on target '%s' has parent %d (%s, %s). Parent: utime=" KERNEL_UINT_FORMAT ", stime=" KERNEL_UINT_FORMAT ", gtime=" KERNEL_UINT_FORMAT ", minflt=" KERNEL_UINT_FORMAT ", majflt=" KERNEL_UINT_FORMAT ", cutime=" KERNEL_UINT_FORMAT ", cstime=" KERNEL_UINT_FORMAT ", cgtime=" KERNEL_UINT_FORMAT ", cminflt=" KERNEL_UINT_FORMAT ", cmajflt=" KERNEL_UINT_FORMAT "", p->pid, p->comm, p->updated?"running":"exited", (p->target)?p->target->name:"UNSET", pp->pid, pp->comm, pp->updated?"running":"exited", pp->utime, pp->stime, pp->gtime, pp->minflt, pp->majflt, pp->cutime, pp->cstime, pp->cgtime, pp->cminflt, pp->cmajflt);
-        }
-        else {
-            p->parent = NULL;
-            netdata_log_error("pid %d %s states parent %d, but the later does not exist.", p->pid, p->comm, p->ppid);
-        }
-    }
-}
-
-// ----------------------------------------------------------------------------
-
-// 1. read all files in /proc
-// 2. for each numeric directory:
-//    i.   read /proc/pid/stat
-//    ii.  read /proc/pid/status
-//    iii. read /proc/pid/io (requires root access)
-//    iii. read the entries in directory /proc/pid/fd (requires root access)
-//         for each entry:
-//         a. find or create a struct file_descriptor
-//         b. cleanup any old/unused file_descriptors
-
-// after all these, some pids may be linked to targets, while others may not
-
-// in case of errors, only 1 every 1000 errors is printed
-// to avoid filling up all disk space
-// if debug is enabled, all errors are printed
-
-#if (ALL_PIDS_ARE_READ_INSTANTLY == 0)
-static int compar_pid(const void *pid1, const void *pid2) {
-
-    struct pid_stat *p1 = all_pids[*((pid_t *)pid1)];
-    struct pid_stat *p2 = all_pids[*((pid_t *)pid2)];
-
-    if(p1->sortlist > p2->sortlist)
-        return -1;
-    else
-        return 1;
-}
-#endif
-
-static int collect_data_for_all_processes(void) {
-    struct pid_stat *p = NULL;
-
-#ifndef __FreeBSD__
-    // clear process state counter
-    memset(proc_state_count, 0, sizeof proc_state_count);
-#else
-    int i, procnum;
-
-    static size_t procbase_size = 0;
-    static struct kinfo_proc *procbase = NULL;
-
-    size_t new_procbase_size;
-
-    int mib[3] = { CTL_KERN, KERN_PROC, KERN_PROC_PROC };
-    if (unlikely(sysctl(mib, 3, NULL, &new_procbase_size, NULL, 0))) {
-        netdata_log_error("sysctl error: Can't get processes data size");
-        return 0;
-    }
-
-    // give it some air for processes that may be started
-    // during this little time.
-    new_procbase_size += 100 * sizeof(struct kinfo_proc);
-
-    // increase the buffer if needed
-    if(new_procbase_size > procbase_size) {
-        procbase_size = new_procbase_size;
-        procbase = reallocz(procbase, procbase_size);
-    }
-
-    // sysctl() gets from new_procbase_size the buffer size
-    // and also returns to it the amount of data filled in
-    new_procbase_size = procbase_size;
-
-    // get the processes from the system
-    if (unlikely(sysctl(mib, 3, procbase, &new_procbase_size, NULL, 0))) {
-        netdata_log_error("sysctl error: Can't get processes data");
-        return 0;
-    }
-
-    // based on the amount of data filled in
-    // calculate the number of processes we got
-    procnum = new_procbase_size / sizeof(struct kinfo_proc);
-
-#endif
-
-    if(all_pids_count) {
-#if (ALL_PIDS_ARE_READ_INSTANTLY == 0)
-        size_t slc = 0;
-#endif
-        for(p = root_of_pids; p ; p = p->next) {
-            p->read             = false; // mark it as not read, so that collect_data_for_pid() will read it
-            p->updated          = false;
-            p->merged           = false;
-            p->children_count   = 0;
-            p->parent           = NULL;
-
-#if (ALL_PIDS_ARE_READ_INSTANTLY == 0)
-            all_pids_sortlist[slc++] = p->pid;
-#endif
-        }
-
-#if (ALL_PIDS_ARE_READ_INSTANTLY == 0)
-        if(unlikely(slc != all_pids_count)) {
-            netdata_log_error("Internal error: I was thinking I had %zu processes in my arrays, but it seems there are %zu.", all_pids_count, slc);
-            all_pids_count = slc;
-        }
-
-        if(include_exited_childs) {
-            // Read parents before childs
-            // This is needed to prevent a situation where
-            // a child is found running, but until we read
-            // its parent, it has exited and its parent
-            // has accumulated its resources.
-
-            qsort((void *)all_pids_sortlist, (size_t)all_pids_count, sizeof(pid_t), compar_pid);
-
-            // we forward read all running processes
-            // collect_data_for_pid() is smart enough,
-            // not to read the same pid twice per iteration
-            for(slc = 0; slc < all_pids_count; slc++) {
-                collect_data_for_pid(all_pids_sortlist[slc], NULL);
-            }
-        }
-#endif
-    }
-
-#ifdef __FreeBSD__
-    for (i = 0 ; i < procnum ; ++i) {
-        pid_t pid = procbase[i].ki_pid;
-        collect_data_for_pid(pid, &procbase[i]);
-    }
-#else
-    static char uptime_filename[FILENAME_MAX + 1] = "";
-    if(*uptime_filename == '\0')
-        snprintfz(uptime_filename, FILENAME_MAX, "%s/proc/uptime", netdata_configured_host_prefix);
-
-    global_uptime = (kernel_uint_t)(uptime_msec(uptime_filename) / MSEC_PER_SEC);
-
-    char dirname[FILENAME_MAX + 1];
-
-    snprintfz(dirname, FILENAME_MAX, "%s/proc", netdata_configured_host_prefix);
-    DIR *dir = opendir(dirname);
-    if(!dir) return 0;
-
-    struct dirent *de = NULL;
-
-    while((de = readdir(dir))) {
-        char *endptr = de->d_name;
-
-        if(unlikely(de->d_type != DT_DIR || de->d_name[0] < '0' || de->d_name[0] > '9'))
-            continue;
-
-        pid_t pid = (pid_t) strtoul(de->d_name, &endptr, 10);
-
-        // make sure we read a valid number
-        if(unlikely(endptr == de->d_name || *endptr != '\0'))
-            continue;
-
-        collect_data_for_pid(pid, NULL);
-    }
-    closedir(dir);
-#endif
-
-    if(!all_pids_count)
-        return 0;
-
-    // we need /proc/stat to normalize the cpu consumption of the exited childs
-    read_global_time();
-
-    // build the process tree
-    link_all_processes_to_their_parents();
-
-    // normally this is done
-    // however we may have processes exited while we collected values
-    // so let's find the exited ones
-    // we do this by collecting the ownership of process
-    // if we manage to get the ownership, the process still runs
-    process_exited_processes();
-    return 1;
 }
 
 // ----------------------------------------------------------------------------
@@ -961,7 +491,6 @@ static inline void post_aggregate_targets(struct target *root) {
 }
 
 static void calculate_netdata_statistics(void) {
-
     apply_apps_groups_targets_inheritance();
 
     zero_all_targets(users_root_target);
