+++ src/collectors/apps.plugin/apps_plugin.h
@@ -6,6 +6,28 @@
 #include "collectors/all.h"
 #include "libnetdata/libnetdata.h"
 
+#ifdef __FreeBSD__
+#include <sys/user.h>
+#endif
+
+// ----------------------------------------------------------------------------
+// per O/S configuration
+
+// the minimum PID of the system
+// this is also the pid of the init process
+#define INIT_PID 1
+
+// if the way apps.plugin will work, will read the entire process list,
+// including the resource utilization of each process, instantly
+// set this to 1
+// when set to 0, apps.plugin builds a sort list of processes, in order
+// to process children processes, before parent processes
+#ifdef __FreeBSD__
+#define ALL_PIDS_ARE_READ_INSTANTLY 1
+#else
+#define ALL_PIDS_ARE_READ_INSTANTLY 0
+#endif
+
 extern bool debug_enabled;
 extern bool enable_guest_charts;
 extern bool enable_detailed_uptime_charts;
@@ -424,6 +446,10 @@ extern int update_every;
 extern unsigned int time_factor;
 extern kernel_uint_t MemTotal;
 
+#if (ALL_PIDS_ARE_READ_INSTANTLY == 0)
+extern pid_t *all_pids_sortlist;
+#endif
+
 #define APPS_PLUGIN_PROCESSES_FUNCTION_DESCRIPTION "Detailed information on the currently running processes."
 
 void function_processes(const char *transaction, char *function,
@@ -491,7 +517,7 @@ int read_proc_pid_io(struct pid_stat *p, void *ptr);
 int read_pid_file_descriptors(struct pid_stat *p, void *ptr);
 int read_global_time(void);
 
-int collect_data_for_pid(pid_t pid, void *ptr);
+int collect_data_for_all_processes(void);
 void cleanup_exited_pids(void);
 
 void clear_pid_fd(struct pid_fd *pfd);
